+++ example/example-app/example-app-cmd-meta/pom.xml
@@ -23,6 +23,18 @@
             <version>${project.version}</version>
         </dependency>
 
+        <dependency>
+            <groupId>com.codingapi.springboot</groupId>
+            <artifactId>example-infra-flow</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.codingapi.springboot</groupId>
+            <artifactId>example-infra-security</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
     </dependencies>
 
 </project>
