--- example/example-application/src/main/java/com/codingapi/example/handler/MyLeaveHandler.java
@@ -1,14 +1,0 @@
-package com.codingapi.example.handler;
-
-import com.codingapi.example.event.UserEvent;
-import com.codingapi.springboot.framework.event.IHandler;
-import org.springframework.stereotype.Component;
-
-@Component
-public class MyLeaveHandler implements IHandler<UserEvent> {
-
-    @Override
-    public void handler(UserEvent event) {
-
-    }
-}
