--- example/example-application/src/main/java/com/codingapi/example/runner/UserRunner.java
@@ -1,19 +1,0 @@
-package com.codingapi.example.runner;
-
-import com.codingapi.example.service.UserService;
-import lombok.AllArgsConstructor;
-import org.springframework.boot.ApplicationArguments;
-import org.springframework.boot.ApplicationRunner;
-import org.springframework.stereotype.Service;
-
-@Service
-@AllArgsConstructor
-public class UserRunner implements ApplicationRunner {
-
-    private final UserService userService;
-
-    @Override
-    public void run(ApplicationArguments args) throws Exception {
-        userService.initAdmin();
-    }
-}
