+++ example/example-interface/src/main/java/com/codingapi/example/handler/LeaveHandler.java
@@ -1,0 +1,23 @@
+package com.codingapi.example.handler;
+
+import com.codingapi.example.domain.leave.entity.Leave;
+import com.codingapi.example.domain.leave.repository.LeaveRepository;
+import com.codingapi.springboot.flow.event.FlowApprovalEvent;
+import com.codingapi.springboot.framework.event.IHandler;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class LeaveHandler implements IHandler<FlowApprovalEvent> {
+
+    private final LeaveRepository leaveRepository;
+
+    @Override
+    public void handler(FlowApprovalEvent event) {
+        if(event.isFinish() && event.match(Leave.class)){
+            Leave leave = (Leave)event.getBindData();
+            leaveRepository.save(leave);
+        }
+    }
+}
