+++ example/example-interface/src/main/java/com/codingapi/example/runner/UserRunner.java
@@ -1,0 +1,19 @@
+package com.codingapi.example.runner;
+
+import com.codingapi.example.domain.user.service.UserService;
+import lombok.AllArgsConstructor;
+import org.springframework.boot.ApplicationArguments;
+import org.springframework.boot.ApplicationRunner;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class UserRunner implements ApplicationRunner {
+
+    private final UserService userService;
+
+    @Override
+    public void run(ApplicationArguments args) throws Exception {
+        userService.initAdmin();
+    }
+}
